Properly instantiate inline module scripts
This moves the latter half of the "internal module script graph fetching
procedure" into "fetch the descendants of a module script", which is
then renamed "fetch the descendants of and instantiate a module script".
Then, the inline script code can properly call this new algorithm to
take care of everything once it has an initial module script prepared
from the inline script source.

Fixes #2544.